228598 / Роли

See merge request grade/v.skriplyuk!26

+++ app/Http/Controllers/ArticlesController.php
@@ -26,6 +26,8 @@ class ArticlesController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Article::class);
+
         $article = $this->articlesRepository->getModel();
         return view('pages.create', compact('article'));
     }
@@ -34,6 +36,8 @@ class ArticlesController extends Controller
         ArticleRequest $request,
         TagsRequest $tagsRequest, CreateArticleServiceContract $createArticleService)
     {
+        $this->authorize('create', Article::class);
+
         $createArticleService->create($request->validated(), $tagsRequest->get('tags'));
 
         return back()->with('success_message', 'Новость успешно создана');
@@ -48,6 +52,9 @@ class ArticlesController extends Controller
     public function edit(string $slug)
     {
         $article = $this->articlesRepository->findBySlug($slug);
+
+        $this->authorize('update', $article);
+
         return view('pages.edit', compact('article'));
     }
 
@@ -57,6 +64,8 @@ class ArticlesController extends Controller
         TagsRequest $tagsRequest,
         UpdateArticleServiceContract $updateArticleService)
     {
+        $this->authorize('update', $this->articlesRepository->findBySlug($slug));
+
         $article =  $updateArticleService->update($slug, $request->validated(), $tagsRequest->get('tags'));
         
         return redirect(route('articles.edit', ['article' => $article]))->with('success_message', 'Новость успешно обновлена');
@@ -64,6 +73,8 @@ class ArticlesController extends Controller
 
     public function destroy(string $slug)
     {
+        $this->authorize('update', $this->articlesRepository->findBySlug($slug));
+        
         $this->articlesRepository->delete($slug);
 
         return redirect(route('articles.index'))->with('success_message', 'Новость удалена');
